Single observable stream for restaurant deals data (#36)
Trying to maintain a reactive pattern for fetching and displaying data..

This means `getRestaurantDeals` acts as a publisher (triggers updates to
_accumulatedDeals), and `accumulatedDeals()` is the subscription point
for consumers.

From now on, view models ("consumers" of the repository data), just need
to have a collector to listen to `accumulatedDeals()` emissions (this
means for the duration of the view models life cycle, we have a stream
waiting for updates from `accumulatedDeals`).

So now when we actually want to update the "store/list" of deals that
ALL view models are looking at (list and map view), call
`getRestaurantDeals()` from anywhere, and we'll add whatever new deals
we get into our `_accumulatedDeals`, and all screens/view models that
are subscribed will get those changes without having to explicitly poll
for them and wait on the result.

To make this pattern extra explicit, `getRestaurantDeals()` no longer
returns the response that it gets.

This means that all view models have the same data, and we only need to
make the call from one place for it to update everywhere.

---------

Co-authored-by: euilooo <89674003+euilooo@users.noreply.github.com>