Pull coords and location permission out of MapViewModel and give it to AppViewModel (#41)
Now that we get the current user location is in the `AppViewModel`
instead of the `mapViewModel`, our `mapViewModel` needs to subscribe to
collect those properties from `AppViewModel` to update its own state. On
a camera event (when we touch the map), we update the appViewModel's
`currentUserLocation`.

Should be no functional changes!! if there are I've done something wrong
D;