Deal model front end (include day time availability restrictions) (#43)
The `RawDeal` object, which reflects what we get from the backend, gets
a `List<Int>?`. If this is `null`, then there are NO restrictions.

```
// means 4-7PM on Monday - Thursday
                        dailyStartTimes = listOf(840, 840, 840, 840, 0, 0, 0),
                        dailyEndTimes = listOf(1080, 1080, 1080, 1080, 0, 0, 0),

// Friday only deal, from 0 - (23 * 60 + 59) minutes
                        dailyStartTimes = listOf(0, 0, 0, 0, 0, 0, 0),
                        dailyEndTimes = listOf(0, 0, 0, 1439, 0, 0),
```

The object that the front end cares about is the `Deal` object. This has
an `activeDayTime` field. The two "raw" arrays of ints gets mapped into
meaningful information (see `mapStartEndTimesToRestriction`)