Force optimistic updates on deal actions (voting and saving) (#54)
Problem: upon taking an action on the vote, the underlying data store
and state (which the UI reflects) is not being updated

Solution: an action on a deal should trigger a fetch to get the
latest-up-to-date copy of the deal, from the source of truth. In our
case this should be the DB.
- I think user vote or saved should return this object (@joycej8
thoughts?)

Since getting this latest-up-to-date copy of the deal could take half a
second, we optimistically update what we have in `accumulatedDeals` and
have the UI (which looks at that) immediately reflect the changes even
if we don't know "for sure" that the vote/save action went through.

This PR adds the optimitic update part (manually changing our global
state to what we think the state will be if the request succeeds). This
is meh good enough maybe for the demo but not actually robust as an
implementation.



https://github.com/user-attachments/assets/19cb8285-de6e-4219-bece-3f6f7793517e